requirements.txt

fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
scikit-learn
nltk
rank_bm25
pypdf
tqdm
python-dotenv